requirements.txt
Flask==3.1.0
SQLAlchemy==2.0.38
psycopg[binary]==3.2.6
PyJWT==2.10.1
python-dotenv==1.0.1
argon2-cffi==23.1.0
pytest==8.3.5
